Happy Birthday.
After looking at your code, I have seen you do not minify your mod scripts so I figured I would post this as is. If there s anything wrong with it, then I don't know what it is because I've checked every line to the point of insanity to make sure it would work as an import.