I unscrewed it.
I don't understand what was going on with the :not tokens but I decided to move `.btn.default` out from the black text block as an exemption and placed `.btn.btn-default` in the white text block. Don't understand what's breaking either; Chaining `:not` tokens shouldn't cause issues as anything which isn't _there_ should remain benign but it doesn't. Oh well.