Because fashion.
I am not sure if this should even be a thing; I was bored, then I made this after experimenting with gradients in ClearVision.

I remain uncertain if most users of these mods would be willing to go through the extra bother of appending their root block. Give this one a pass if you think that's a bit much to expect.